apps/api/pyproject.toml
[project]
name = "hearsay-api"
version = "0.1.0"
description = "PDF Study Reader API — OCR + TTS backend"
requires-python = ">=3.12"
dependencies = [
    "convex>=0.7.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "nltk>=3.9.3",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
    "surya-ocr>=0.13.1",
    "transformers>=4.56.1,<5.0.0",
    "requests>=2.32.5",
    "anthropic>=0.40.0",
]

[tool.uv]
package = false
